Update build and run instructions to be able to pick back up
The blocker issue for this repository has been cleared so I want to pick the work back up but I am finding I did not do a good enough job of capturing the build instructions and as I reverse engineer them I am finding that the CRA build command is not working with the latest Node LTS so I will need to crack that issue next. In the meanwhile, this is the updated instructions as far as I have been able to recover them for now.