Replace the old POCs with a working solution based on the repro repository
I ended up basically implementing the core of what I wanted to prototype here in the related repository https://github.com/TomasHubelbauer/babel-sourcemap where I was demonstrating a source-map problem to the Babel folks.

Nevertheless, I redid the same here and then added a few sprinkles on top. It ended up behaving differently anyway, see the last TODO in the README as of this commit.

I added a few bells and whistles in this copy since this is to be the real deal once.

I have decoupled this repository from the cra submodule one because I found a better way to scaffold CRA transparently using package scripts. I removed that repository off of GitHub and made a TODO here to add a GH Actions workflow to rebuild on new CRA versions automatically.

I am not yet sending in the contents of docs, I will send that in a separate commit.